# api/adminapi/openapi.yaml
openapi: 3.0.2
info:
  title: Federation Admin
  version: 0.1.0
  description: >-
    This API enables management of federated entity data and metadata within an OpenID Federation
    environment.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8765
    description: Local development server
...
paths:
...
  /api/v1/admin/federation-endpoints:
    summary: Manage federation endpoints.
    description: List and create federation endpoint configurations.
    get:
      tags:
        - Federation Endpoints
      responses:
        '200':
          description: List of federation endpoints.
      operationId: listFederationEndpoints
      summary: List federation endpoints
    post:
      tags:
        - Federation Endpoints
      responses:
        '201':
          description: Federation endpoint created.
        '409':
          description: Federation endpoint already exists.
      operationId: createFederationEndpoint
      summary: Create a federation endpoint
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddFederationEndpoint'
  /api/v1/admin/federation-endpoints/{type}:
    summary: Manage a single federation endpoint by type.
    description: Get, update, or delete a federation endpoint by type.
    get:
      tags:
        - Federation Endpoints
      responses:
        '200':
          description: Federation endpoint details.
        '404':
          description: Federation endpoint not found.
      operationId: getFederationEndpoint
      summary: Get a federation endpoint
    put:
      tags:
        - Federation Endpoints
      responses:
        '200':
          description: Federation endpoint updated.
        '404':
          description: Federation endpoint not found.
      operationId: updateFederationEndpoint
      summary: Update a federation endpoint
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddFederationEndpoint'
    delete:
      tags:
        - Federation Endpoints
      responses:
        '204':
          description: Federation endpoint deleted.
        '404':
          description: Federation endpoint not found.
      operationId: deleteFederationEndpoint
      summary: Delete a federation endpoint
    parameters:
      - name: type
        description: The endpoint type (fetch, list, resolve, trust_mark, etc.).
        schema:
          type: string
        in: path
        required: true
  /api/v1/admin/federation-endpoints/{type}/auth-trust-anchors:
    summary: Set auth trust anchors for an endpoint.
    description: Replace the trust anchors used for endpoint authentication.
    put:
      tags:
        - Federation Endpoints
      responses:
        '200':
          description: Auth trust anchors updated.
        '404':
          description: Federation endpoint or trust anchor not found.
      operationId: setFederationEndpointAuthTrustAnchors
      summary: Set auth trust anchors for a federation endpoint
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: integer
    parameters:
      - name: type
        description: The endpoint type.
        schema:
          type: string
        in: path
        required: true
components:
  schemas:
...
    AddFederationEndpoint:
      description: Request payload to create/update a federation endpoint.
      type: object
      required:
        - type
      properties:
        type:
          type: string
          description: Endpoint type (fetch, list, resolve, trust_mark, trust_mark_status, trust_mark_listing, historical_keys, enroll, enroll_request, trust_mark_request, entity_collection).
        path:
          type: string
          description: Internal path for the endpoint (nullable; null = disabled).
        url:
          type: string
          description: External URL for the endpoint (nullable; auto-derived from path if null).
        auth_enabled:
          type: boolean
          description: Enable private_key_jwt authentication for this endpoint.
          default: false
        config:
          type: string
          description: JSON blob with type-specific configuration.
        auth_trust_anchor_ids:
          type: array
          items:
            type: integer
          description: Trust anchor IDs used for endpoint authentication.
...